Make dangling_container stuff safe to run again
The existing logic was:
 - Check for the presence of a `com.hashicorp.nomad.alloc_id` label.
   If found, it's "ours". Otherwise, continue.
 - If the container lacks any of `/alloc`, `/local`, or `/secrets` or
   its name doesn't match something like "SomeString-SomeUUID", it's
   NOT ours. Otherwise, we guess it's ours.

The first step was supposed to suffice, but the second step exists to
handle containers from pre-0.10 Nomad. I think worrying about containers
from before 2019 is over the top, so I removed that logic from this
plugin, but it remains the behavior of the internal Docker plugin, so
we need to take it into account.

I change the label we assign from `com.hashicorp.nomad.alloc_id` to
`com.hashicorp.nomad.docker-ext.alloc_id`. This keeps our noses out of
the internal plugin's business, and vice versa in the first step.

We don't want to get rid of the mounts, so we need to avoid the
"SomeString-SomeUUID" check. To do so, I add the plugin name
("docker-ext") to the container name.