Adjust for building/running as external module
Apart from the small changes to the build system, some work had to
be done to avoid Nomad's dangling container logic causing interference
between the built-in docker driver and this one.

The existing logic was:
1. Check for the presence of a `com.hashicorp.nomad.alloc_id` label.
   If found, it's "ours". Otherwise, continue.
2. If the container lacks any of `/alloc`, `/local`, or `/secrets` or
   its name doesn't match something like "SomeString-SomeUUID", it's
   NOT ours.
3. Otherwise, we guess it's ours.

Step 1 was supposed to suffice, but the second step exists to handle
containers from pre-0.10 Nomad. I think worrying about containers
from before 2019 is over the top, so I removed that logic from this
plugin, but it remains the behavior of the internal Docker plugin, so
we need to take it into account.

I change the label we assign from `com.hashicorp.nomad.alloc_id` to
`com.hashicorp.nomad.docker-ext.alloc_id`. This keeps our noses out of
the internal plugin's business, and vice versa in the first step.

We're obviously going to have the same mounts, so to evade the built-in
driver's detection in step 2, I added the plugin name ("docker-ext")
to the container name.